add relationships between tag and user, validations on tag controller

+++ api/app/Controllers/Http/TagController.js
@@ -5,13 +5,14 @@ const { validateAll } = use('Validator')
 const Tag = use('App/Models/Tag')
 
 class TagController {
-  async index ({ response }) {
-    const tags = await Tag.all()
+  async index ({ auth }) {
+    const tags = await auth.user.tags().fetch()
 
     return tags
   }
 
-  async store ({ request, response }) {
+  async store ({ auth, request }) {
+    const { id } = auth.user
     const data = request.only(['name'])
 
     const validation = await validateAll(data, {
@@ -22,18 +23,24 @@ class TagController {
       return validation.messages()
     }
 
+    data.user_id = id
+
     const tag = await Tag.create(data)
 
     return tag
   }
 
-  async show ({ params, response }) {
+  async show ({ auth, params, response }) {
     const tag = await Tag.findOrFail(params.id)
+
+    if (tag.user_id !== auth.user.id) {
+      return response.unauthorized({ error: 'Not authorized' })
+    }
 
     return tag
   }
 
-  async update ({ params, request, response }) {
+  async update ({ auth, params, request, response }) {
     const tag = await Tag.findOrFail(params.id)
     const data = request.only(['name'])
 
@@ -43,6 +50,10 @@ class TagController {
 
     if (validation.fails()) {
       return validation.messages()
+    }
+
+    if (tag.user_id !== auth.user.id) {
+      return response.unauthorized({ error: 'Not authorized.' })
     }
 
     tag.merge(data)
@@ -51,8 +62,12 @@ class TagController {
     return tag
   }
 
-  async delete ({ params, response }) {
+  async destroy ({ auth, params, response }) {
     const tag = await Tag.findOrFail(params.id)
+
+    if (tag.user_id !== auth.user.id) {
+      return response.unauthorized({ error: 'Not authorized.' })
+    }
 
     await tag.delete()
